requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-environ>=0.9.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
requests == 2.28.1